+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [Unreleased]
+- Push manifest keys are now regular expression patterns instead of exact paths.
 
 ## [0.9.0] 2017-08-23
 - Add `--bot-proxy` flag to proxy requests from bots through [Rendertron](https://github.com/GoogleChrome/rendertron).